+++ Changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [UNRELEASED]
 ### Added
  - Possibility to add custom validation message.
+ - Possibility to add shortcuts.
 
 ## 2.24.0
 ### Changed